move lambda resource providers to proper module
