+++ localstack/services/lambda_/resource_providers/lambda_alias.schema.json
